docker/requirements.txt
numpy==1.26.4
opencv-python==4.9.0.80
livekit==0.11.1
python-socketio==5.11.2
aiohttp==3.9.5
requests==2.32.3
websocket-client==1.8.0
